@@ -15,18 +15,18 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.3",
-    "@angular/animations": "^6.1.10",
-    "@angular/cdk": "^7.0.3",
-    "@angular/common": "^6.0.0-rc.1",
-    "@angular/core": "^6.0.0-rc.1",
-    "@angular/forms": "^6.0.0-rc.1",
-    "@angular/http": "^6.0.0-rc.1",
+    "@angular/animations": "^7.1.1",
+    "@angular/cdk": "^7.1.1",
+    "@angular/common": "^7.1.1",
+    "@angular/core": "^7.1.1",
+    "@angular/forms": "^7.1.1",
+    "@angular/http": "^7.1.1",
     "@angular/material": "^7.0.3",
-    "@angular/platform-browser": "^6.0.0-rc.1",
-    "@angular/platform-browser-dynamic": "^6.0.0-rc.1",
-    "@angular/platform-server": "^6.0.0-rc.1",
-    "@angular/platform-webworker": "^7.1.0",
-    "@angular/router": "^6.0.0-rc.1",
+    "@angular/platform-browser": "^7.1.1",
+    "@angular/platform-browser-dynamic": "^7.1.1",
+    "@angular/platform-server": "^7.1.1",
+    "@angular/platform-webworker": "^7.1.1",
+    "@angular/router": "^7.1.1",
     "@auth0/angular-jwt": "^2.0.0",
     "angular-bootstrap-md": "6.2.6",
     "angular-in-memory-web-api": "^0.7.0",
@@ -42,15 +42,16 @@
     "ngx-mask": "^7.0.1",
     "rxjs": "^6.0.0-rc.1",
     "socket.io": "^2.1.1",
+    "tslib": "^1.9.0",
     "web-animations-js": "^2.3.1",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.5.7",
-    "@angular/cli": "^6.2.7",
-    "@angular/compiler": "^6.0.0",
-    "@angular/compiler-cli": "^6.0.0-rc.1",
-    "@angular/language-service": "^6.0.0-rc.1",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "^7.1.0",
+    "@angular/compiler": "^7.1.1",
+    "@angular/compiler-cli": "^7.1.1",
+    "@angular/language-service": "^7.1.1",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -66,6 +67,6 @@
     "protractor": "^5.4.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "~3.1.6"
   }
 }
